Move core class struct and methods to separate file
